# Back/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/auth/register:
    post:
      summary: 'Register User'
      operationId: registerUser
      description: 'Register new user into website'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 11
                    name: Abdulrahman
                    email: abouelez10@gmail.com
                    is_instructor: 0
                    is_admin: 0
                  access_token: 1|lNZjX8LiCgM9HfesupUAw0jBWnYnu9mixpOK5t24
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 11
                      name:
                        type: string
                        example: Abdulrahman
                      email:
                        type: string
                        example: abouelez10@gmail.com
                      is_instructor:
                        type: integer
                        example: 0
                      is_admin:
                        type: integer
                        example: 0
                  access_token:
                    type: string
                    example: 1|lNZjX8LiCgM9HfesupUAw0jBWnYnu9mixpOK5t24
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The name field is required. (and 2 more errors)'
                  errors:
                    name:
                      - 'The name field is required.'
                    email:
                      - 'The email has already been taken.'
                    password:
                      - 'The password field confirmation does not match.'
                properties:
                  message:
                    type: string
                    example: 'The name field is required. (and 2 more errors)'
                  errors:
                    type: object
                    properties:
                      name:
                        type: array
                        example:
                          - 'The name field is required.'
                        items:
                          type: string
                      email:
                        type: array
                        example:
                          - 'The email has already been taken.'
                        items:
                          type: string
                      password:
                        type: array
                        example:
                          - 'The password field confirmation does not match.'
                        items:
                          type: string
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Your full name.'
                  example: Abdulrahman
                email:
                  type: string
                  description: 'Your email.'
                  example: Abouelez10@gmail.com
                password:
                  type: string
                  description: 'Your Password.'
                  example: password
                password_confirmation:
                  type: string
                  description: 'Confirm password.'
                  example: password
              required:
                - name
                - email
                - password_confirmation
      security: []
  /api/auth/login:
    post:
      summary: Login
      operationId: login
      description: 'Login to website'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 11
                    name: Abdulrahman
                    email: abouelez10@gmail.com
                    is_instructor: 0
                    is_admin: 0
                  access_token: 5|AvgBOMYXI0hFXX2yWM5oxBiJD1BVEUXis1flm3GI
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 11
                      name:
                        type: string
                        example: Abdulrahman
                      email:
                        type: string
                        example: abouelez10@gmail.com
                      is_instructor:
                        type: integer
                        example: 0
                      is_admin:
                        type: integer
                        example: 0
                  access_token:
                    type: string
                    example: 5|AvgBOMYXI0hFXX2yWM5oxBiJD1BVEUXis1flm3GI
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The provided credentials are incorrect.'
                properties:
                  message:
                    type: string
                    example: 'The provided credentials are incorrect.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Your email.'
                  example: bernie59@example.org
                password:
                  type: string
                  description: 'Your Password.'
                  example: jOWTq>n$3yse@L
              required:
                - email
      security: []
...
  '/api/add-to-cart/{course_id}':
    get:
      summary: 'Add To Cart'
      operationId: addToCart
      description: 'Add Course To Cart'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Course added to cart'
                  course:
                    id: 2
                    title: 'Rem nostrum dolores tempore est veniam occaecati.'
                    description: 'Et et praesentium enim nemo sunt placeat et. Tempora earum accusantium sint porro. Voluptas facilis natus et eum ut nemo quia. Numquam sint provident nemo sit natus voluptatem atque.'
                    thumbnail: 'https://via.placeholder.com/640x480.png/0099ee?text=ad'
                    price: '938.70'
                    published: 1
                    instructor:
                      id: 5
                      name: 'Wilford Skiles'
                      email: rafaela.wilderman@sd.org
                      avatar: 'https://via.placeholder.com/640x480.png/004444?text=similique'
                      is_instructor: 1
                      is_admin: 0
                    rate: '0.0'
                    total_enrollments: 0
                    your_review: null
                properties:
                  message:
                    type: string
                    example: 'Course added to cart'
                  course:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 2
                      title:
                        type: string
                        example: 'Rem nostrum dolores tempore est veniam occaecati.'
                      description:
                        type: string
                        example: 'Et et praesentium enim nemo sunt placeat et. Tempora earum accusantium sint porro. Voluptas facilis natus et eum ut nemo quia. Numquam sint provident nemo sit natus voluptatem atque.'
                      thumbnail:
                        type: string
                        example: 'https://via.placeholder.com/640x480.png/0099ee?text=ad'
                      price:
                        type: string
                        example: '938.70'
                      published:
                        type: integer
                        example: 1
                      instructor:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 5
                          name:
                            type: string
                            example: 'Wilford Skiles'
                          email:
                            type: string
                            example: rafaela.wilderman@sd.org
                          avatar:
                            type: string
                            example: 'https://via.placeholder.com/640x480.png/004444?text=similique'
                          is_instructor:
                            type: integer
                            example: 1
                          is_admin:
                            type: integer
                            example: 0
                      rate:
                        type: string
                        example: '0.0'
                      total_enrollments:
                        type: integer
                        example: 0
                      your_review:
                        type: string
                        example: null
      tags:
        - Cart
    parameters:
      -
        in: path
        name: course_id
        description: 'The ID of the course.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/remove-from-cart/{course_id}':
    delete:
      summary: Remove
      operationId: remove
      description: 'Remove Course From Cart'
      parameters: []
      responses:
        204:
          description: ''
      tags:
        - Cart
    parameters:
      -
        in: path
        name: course_id
        description: 'The ID of the course.'
        example: 1
        required: true
        schema:
          type: integer
  /api/categories:
    post:
      summary: 'Create Category'
      operationId: createCategory
      description: 'Store a newly created Category in storage.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 16
                    name: new
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 16
                      name:
                        type: string
                        example: new
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The name field is required.'
                  errors:
                    name:
                      - 'The name field is required.'
                properties:
                  message:
                    type: string
                    example: 'The name field is required.'
                  errors:
                    type: object
                    properties:
                      name:
                        type: array
                        example:
                          - 'The name field is required.'
                        items:
                          type: string
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Name of new category.'
                  example: Development
                parent_id:
                  type: integer
                  description: ''
                  example: 5
              required:
                - name
    get:
      summary: 'Get categories'
      operationId: getCategories
      description: 'List all categories'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name: 'Officiis nostrum rerum facere reprehenderit ipsum reprehenderit.'
                      parent: null
                      children: []
                    -
                      id: 2
                      name: 'Et repudiandae quam sit ratione.'
                      parent: null
                      children: []
                    -
                      id: 3
                      name: 'Quam fugiat expedita fuga id inventore quisquam.'
                      parent: null
                      children: []
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name: 'Officiis nostrum rerum facere reprehenderit ipsum reprehenderit.'
                        parent: null
                        children: []
                      -
                        id: 2
                        name: 'Et repudiandae quam sit ratione.'
                        parent: null
                        children: []
                      -
                        id: 3
                        name: 'Quam fugiat expedita fuga id inventore quisquam.'
                        parent: null
                        children: []
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: 'Officiis nostrum rerum facere reprehenderit ipsum reprehenderit.'
                        parent:
                          type: string
                          example: null
                        children:
                          type: array
                          example: []
      tags:
        - Categories
      security: []
...
  /api/comments:
    post:
      summary: 'Add Comment'
      operationId: addComment
      description: 'Comment on course.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'comment published'
                  comment:
                    id: 1
                    comment_text: good
                    user:
                      id: 11
                      name: Abdulrahman
                      email: abouelez10@gmail.com
                      is_instructor: 1
                      is_admin: 1
                properties:
                  message:
                    type: string
                    example: 'comment published'
                  comment:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      comment_text:
                        type: string
                        example: good
                      user:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 11
                          name:
                            type: string
                            example: Abdulrahman
                          email:
                            type: string
                            example: abouelez10@gmail.com
                          is_instructor:
                            type: integer
                            example: 1
                          is_admin:
                            type: integer
                            example: 1
        422:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: '"message":"The course id field is required. (and 1 more error)","errors":{"course_id":["The course id field is required."],"comment_text":["The comment text field is required."]}}'
      tags:
        - Comments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                course_id:
                  type: integer
                  description: required.
                  example: 2
                comment_text:
                  type: string
                  description: 'Your comment.'
                  example: 'Excellent course'
              required:
                - comment_text
...
  '/api/verify-email/{id}/{hash}':
    get:
      summary: 'Verify Email'
      operationId: verifyEmail
      description: 'Verify email for user'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Email Already Verified'
                properties:
                  message:
                    type: string
                    example: 'Email Already Verified'
      tags:
        - 'Email Verification'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the verify email.'
        example: quisquam
        required: true
        schema:
          type: string
      -
        in: path
        name: hash
        description: ''
        example: consequatur
        required: true
        schema:
          type: string
  /api/enroll:
    post:
      summary: Payment
      operationId: payment
      description: 'Payment Process'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Enrollment successful'
                properties:
                  message:
                    type: string
                    example: 'Enrollment successful'
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: "No such PaymentMethod: 'pm_1NS6eqGEPWFHPllANwhUzh'"
                properties:
                  message:
                    type: string
                    example: "No such PaymentMethod: 'pm_1NS6eqGEPWFHPllANwhUzh'"
      tags:
        - 'Enrollment And Payment'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                course_ids:
                  type: object
                  description: 'Array of courses ids.'
                  example:
                    - 1
                    - 10
                  properties: []
                payment_amount:
                  type: number
                  description: 'Amount needed to pay.'
                  example: 100.0
                payment_id:
                  type: string
                  description: 'Payment method id that generated from stripe. Example ='
                  example: nihil
              required:
                - course_ids
                - payment_amount
                - payment_id
  /api/lectures:
    get:
      summary: 'Get Lectures'
      operationId: getLectures
      description: 'Display a listing of the resource.'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Lectures
    post:
      summary: 'Create Lecture'
      operationId: createLecture
      description: 'Create a newly lecture within a course.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 2
                    title: 'new lec'
                    type: video
                    content: /storage/content/courses/course1/section1/lecture2/lecture.mkv
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 2
                      title:
                        type: string
                        example: 'new lec'
                      type:
                        type: string
                        example: video
                      content:
                        type: string
                        example: /storage/content/courses/course1/section1/lecture2/lecture.mkv
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The title field is required. (and 3 more errors)'
                  errors:
                    title:
                      - 'The title field is required.'
                    type:
                      - 'The type field is required.'
                    section_id:
                      - 'The section id field is required.'
                    content:
                      - 'The content field is required.'
                properties:
                  message:
                    type: string
                    example: 'The title field is required. (and 3 more errors)'
                  errors:
                    type: object
                    properties:
                      title:
                        type: array
                        example:
                          - 'The title field is required.'
                        items:
                          type: string
                      type:
                        type: array
                        example:
                          - 'The type field is required.'
                        items:
                          type: string
                      section_id:
                        type: array
                        example:
                          - 'The section id field is required.'
                        items:
                          type: string
                      content:
                        type: array
                        example:
                          - 'The content field is required.'
                        items:
                          type: string
      tags:
        - Lectures
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'Title of lecture.'
                  example: 'lec 1'
                type:
                  type: string
                  description: '[video or text].'
                  example: video
                section_id:
                  type: integer
                  description: require.
                  example: 1
                content:
                  type: string
                  format: binary
                  description: 'Content of lecture'
              required:
                - title
                - type
...
  '/api/notifications/{id}/mark-read':
    get:
      summary: 'Mark as read'
      operationId: markAsRead
      description: 'Mark notification as read'
      parameters:
        -
          in: query
          name: id
          description: ''
          example: debitis
          required: true
          schema:
            type: string
            description: ''
            example: debitis
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Notification marked as read'
                properties:
                  message:
                    type: string
                    example: 'Notification marked as read'
      tags:
        - Notification
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the notification.'
        example: necessitatibus
        required: true
        schema:
          type: string
...
  /api/active-admin-role:
    get:
      summary: 'Admin Role'
      operationId: adminRole
      description: 'Make user admin'
      parameters:
        -
          in: query
          name: email
          description: 'email of user to be admin.'
          example: test@t.com
          required: true
          schema:
            type: string
            description: 'email of user to be admin.'
            example: test@t.com
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  notification: 'specified user  now instructor'
                properties:
                  notification:
                    type: string
                    example: 'specified user  now instructor'
      tags:
        - 'User Management'
...
  '/api/remove-from-wishlist/{course_id}':
    delete:
      summary: Remove
      operationId: remove
      description: 'Remove Course From Wishlist'
      parameters: []
      responses:
        204:
          description: ''
      tags:
        - wishlist
    parameters:
      -
        in: path
        name: course_id
        description: 'The ID of the course.'
        example: 1
        required: true
        schema:
          type: integer
...